+++ CHANGELOG.md
@@ -11,9 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Autosave instructions
 - Editable instructions
 - Support for the `outputPanels` attribute in the `PyodideRunner` (#1157)
-- Downloading project instructions (#116
+- Downloading project instructions (#1160)
 - Show instructions option in sidebar if instructions are editable (#1164)
 - Open instructions panel by default if instructions are editable (#1164)
+- Instructions empty state to show when instructions are editable (#1165)
 
 ### Changed